PlanGenerator.createOfflineDql() method
